refactor: move scope utility functions to internal/authutil package
