+++ auth/shared.go
@@ -8,14 +8,11 @@ package auth
 
 import (
 	"context"
-	"maps"
 	"net/http"
 	"net/url"
-	"slices"
 	"strings"
 
 	"github.com/modelcontextprotocol/go-sdk/oauthex"
-	"golang.org/x/oauth2"
 )
 
 // GetAuthServerMetadata fetches authorization server metadata for the given issuer URL.
@@ -69,28 +66,4 @@ func authorizationServerMetadataURLs(issuerURL string) []string {
 	urls = append(urls, baseURL.String())
 
 	return urls
-}
-
-// UnionScopes returns the union of the existing and challenged scope sets.
-// It is used during step-up authorization to accumulate scopes across
-// authorization rounds (SEP-2350).
-func UnionScopes(existing, challenged []string) []string {
-	combined := make(map[string]struct{})
-	for _, s := range existing {
-		combined[s] = struct{}{}
-	}
-	for _, s := range challenged {
-		combined[s] = struct{}{}
-	}
-	return slices.Collect(maps.Keys(combined))
-}
-
-// ScopesFromToken extracts the granted scopes from an OAuth2 token response.
-// Per RFC 6749 §5.1, the scope parameter is optional; returns nil if absent.
-func ScopesFromToken(token *oauth2.Token) []string {
-	scope, ok := token.Extra("scope").(string)
-	if !ok {
-		return nil
-	}
-	return strings.Fields(scope)
 }
