+++ auth/shared_test.go
@@ -8,8 +8,6 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/modelcontextprotocol/go-sdk/internal/oauthtest"
 )
 
@@ -95,67 +93,6 @@ func TestGetAuthServerMetadata(t *testing.T) {
 			}
 			if got.Issuer != issuerURL {
 				t.Errorf("GetAuthServerMetadata() issuer = %q, want %q", got.Issuer, issuerURL)
-			}
-		})
-	}
-}
-
-func TestUnionScopes(t *testing.T) {
-	tests := []struct {
-		name       string
-		existing   []string
-		challenged []string
-		want       []string
-	}{
-		{
-			name:       "both empty",
-			existing:   nil,
-			challenged: nil,
-			want:       nil,
-		},
-		{
-			name:       "existing only",
-			existing:   []string{"read"},
-			challenged: nil,
-			want:       []string{"read"},
-		},
-		{
-			name:       "challenged only",
-			existing:   nil,
-			challenged: []string{"write"},
-			want:       []string{"write"},
-		},
-		{
-			name:       "disjoint scopes",
-			existing:   []string{"read"},
-			challenged: []string{"write"},
-			want:       []string{"read", "write"},
-		},
-		{
-			name:       "overlapping scopes",
-			existing:   []string{"read", "write"},
-			challenged: []string{"write", "admin"},
-			want:       []string{"read", "write", "admin"},
-		},
-		{
-			name:       "identical scopes",
-			existing:   []string{"read", "write"},
-			challenged: []string{"read", "write"},
-			want:       []string{"read", "write"},
-		},
-		{
-			name:       "mixed scopes",
-			existing:   []string{"b", "a"},
-			challenged: []string{"c", "a"},
-			want:       []string{"a", "b", "c"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := UnionScopes(tt.existing, tt.challenged)
-			if diff := cmp.Diff(tt.want, got, cmpopts.SortSlices(func(a, b string) bool { return a < b })); diff != "" {
-				t.Errorf("UnionScopes() mismatch (-want +got):\n%s", diff)
 			}
 		})
 	}
